+++ apps/native-component-list/src/navigation/ExpoComponentsStackNavigator.tsx
@@ -389,6 +389,12 @@ export const Screens: ScreenConfig[] = [
   },
   {
     getComponent() {
+      return optionalRequire(() => require('../screens/UIScreen'));
+    },
+    name: 'Expo UI',
+  },
+  {
+    getComponent() {
       return optionalRequire(() => require('../screens/Screens'));
     },
     name: 'Screens',