+++ apps/native-component-list/src/screens/ExpoComponentsScreen.tsx
@@ -44,6 +44,7 @@ const screens = [
   'Touchables',
   'Video (expo-av)',
   'Video (expo-video)',
+  'Expo UI',
   'PagerView',
   'WebView',
 ];