+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/views/ExpoView.kt
@@ -53,6 +53,9 @@ abstract class ExpoView(
     // When the border radius is set, we need to clip the content to the padding box.
     // This is because the border radius is applied to the background drawable, not the view itself.
     // It is the same behavior as in React Native.
+    if (!clipToPadding) {
+      return
+    }
     BackgroundStyleApplicator.clipToPaddingBox(this, canvas)
     super.dispatchDraw(canvas)
   }