+++ packages/expo-ui/src/ExpoUIModule.ts
@@ -1,0 +1,5 @@
+import { NativeModule, requireNativeModule } from 'expo';
+
+declare class ExpoUIModule extends NativeModule {}
+
+export default requireNativeModule<ExpoUIModule>('ExpoUI');